Refactoring: resturct files and dirs
