+++ .size-limit
@@ -1,7 +1,7 @@
 [
 	{
 		"path": "package/index.js",
-		"limit": "15.7 KB",
+		"limit": "16.1 KB",
 		"webpack": false
 	},
 	{
