+++ src/utils/index.js
@@ -1,0 +1,3 @@
+export * from './loader';
+export * from './mimeTypes';
+export * from './module';
